Rust - Improve utils and unzipping encode results
