+++ tokenizers/src/utils/truncation.rs
@@ -1,26 +1,10 @@
-use crate::tokenizer::{Encoding, PaddingDirection, Result};
-use rayon::prelude::*;
+use crate::tokenizer::{Encoding, Result};
 
 #[derive(Debug, Clone)]
 pub struct TruncationParams {
     pub max_length: usize,
     pub strategy: TruncationStrategy,
     pub stride: usize,
-}
-
-#[derive(Debug, Clone)]
-pub struct PaddingParams {
-    pub strategy: PaddingStrategy,
-    pub direction: PaddingDirection,
-    pub pad_id: u32,
-    pub pad_type_id: u32,
-    pub pad_token: String,
-}
-
-#[derive(Debug, Clone)]
-pub enum PaddingStrategy {
-    BatchLongest,
-    Fixed(usize),
 }
 
 #[derive(Debug)]
@@ -117,34 +101,4 @@ pub fn truncate_encodings(
     }
 
     Ok((encoding, pair_encoding))
-}
-
-pub fn pad_encodings(
-    mut encodings: Vec<Encoding>,
-    params: &PaddingParams,
-) -> Result<Vec<Encoding>> {
-    if encodings.is_empty() {
-        return Ok(encodings);
-    }
-
-    let pad_length = match params.strategy {
-        PaddingStrategy::Fixed(size) => size,
-        PaddingStrategy::BatchLongest => encodings
-            .par_iter()
-            .map(|e| e.get_ids().len())
-            .max()
-            .unwrap(),
-    };
-
-    encodings.par_iter_mut().for_each(|encoding| {
-        encoding.pad(
-            pad_length,
-            params.pad_id,
-            params.pad_type_id,
-            &params.pad_token,
-            params.direction,
-        )
-    });
-
-    Ok(encodings)
 }
